Reverse unsafe/bit hacks for more sensible struct

+++ algorithms_test.go
@@ -160,16 +160,16 @@ func constructTree(K int) karytree.Node {
 	var curr *karytree.Node
 	curr = &tree
 
-	for k := uint16(0); k < uint16(K); k++ {
+	for k := uint(0); k < uint(K); k++ {
 		child := karytree.NewNode(key)
 		key++
 		curr.SetNthChild(k, &child)
-		for l := uint16(0); l < uint16(K); l++ {
+		for l := uint(0); l < uint(K); l++ {
 			grandchild := karytree.NewNode(key)
 			key++
 			nth := curr.NthChild(k)
 			nth.SetNthChild(l, &grandchild)
-			for m := uint16(0); m < uint16(K); m++ {
+			for m := uint(0); m < uint(K); m++ {
 				greatgrandchild := karytree.NewNode(key)
 				key++
 
@@ -193,14 +193,14 @@ func constructTreeSparse(K int) karytree.Node {
 	var curr *karytree.Node
 	curr = &tree
 
-	for i := uint16(0); i < uint16(K); i++ {
+	for i := uint(0); i < uint(K); i++ {
 		if i%2 == 0 {
 			child := karytree.NewNode(key)
 			key++
 
 			// fill even children
 			curr.SetNthChild(i, &child)
-			for j := uint16(0); j < uint16(K); j++ {
+			for j := uint(0); j < uint(K); j++ {
 				if j%2 != 0 {
 					grandchild := karytree.NewNode(key)
 					key++
@@ -208,7 +208,7 @@ func constructTreeSparse(K int) karytree.Node {
 
 					// fill odd grandchildren
 					ith.SetNthChild(j, &grandchild)
-					for k := uint16(0); k < uint16(K); k++ {
+					for k := uint(0); k < uint(K); k++ {
 						if k%2 == 0 {
 							// fill even great grandchildren
 							greatgrandchild := karytree.NewNode(key)
